updated Env int()

+++ Strawry/env.py
@@ -23,7 +23,8 @@ class Env(object):
         x3=self.cap(self.co2, 300, 3000, 5)
         x4=self.light
         x5=self.watp
-        return x1*5*5*2*2 + x2*5*5*2*2 + x3*2*2 + x4*2 +x5
+        return int(x1*5*5*2*2 + x2*5*5*2*2 + x3*2*2 + x4*2 +x5)
+        
     
     def reset(self, temp=25, humi=80, co2=200, light=0, watp=0):        
         self.temp=temp
